+++ wazo_calld/plugins/applications/resources.py
@@ -117,6 +117,16 @@ class ApplicationCallMuteStopList(_BaseResource):
         return '', 204
 
 
+class ApplicationCallAnswer(_BaseResource):
+
+    @required_acl('calld.applications.{application_uuid}.calls.{call_id}.answer.update')
+    def put(self, application_uuid, call_id):
+        application = self._service.get_application(application_uuid)
+        self._service.get_call_id(application, call_id)
+        self._service.call_answer(application, call_id)
+        return '', 204
+
+
 class ApplicationCallPlaybackList(_BaseResource):
 
     @required_acl('calld.applications.{application_uuid}.calls.{call_id}.playbacks.create')